Added production time to article; Order renamed to ArticleOrder(because SQL conflict to word "order"; Add Order list page, multiselect for articles; Add ArticleInOrder table for article status in each order.

+++ src/main/java/org/jewel/db/ArticleRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 @Repository
 @RepositoryRestResource(collectionResourceRel = "articles", itemResourceRel = "articles", path = "articles")
-public interface ArticleRepository extends PagingAndSortingRepository<Article, Integer> {
+public interface ArticleRepository extends PagingAndSortingRepository<Article, Long> {
     List<Article> findArticlesByMetalType(MetalType metalType);
 
     List<Article> findArticlesByArticleName(String articleName);
